ISSUE #45 and #47
1. Changed every method that uses DBManager-methods in these classes:
--> ActivityManager.java
--> ActivityManagerTest.java
--> SPManager.java
--> SPManagerTest.java
--> UserManager.java
--> UserManagerTest.java
--> FxFilteringController.java
--> MainScreenController.java
--> UserScreenController.java

2. Changed the methods getRecommendedDailyActivity(), getShareValue(),
and setShareValue() in HttpCommunication.java so that these methods now
takes in the attribute "String tableName". This makes it possible to
test the methods with the use of the test-tables.

3. Created a testclass called HttpCommunicationTestIT.java in the web
server module that so far tests getShareValue() and setShareValue() in
HttpCommunication.java.

4. Changed the doGet-method in getRecommendedDailyActivityServlet.java
and ShareDataServlet.java so that the tableName attribute is passed on
to the query.

5. Removed testGetValueOfSharing.java from UserScreenTest.java.

6. Added tests testAddUser.java and testDeleteUser.java in
UserManagerTest.java.

7. Commented out the testGetRecommendedDailyActivity in
HttpCommunicationTest.java because the Maven Install failed, even though
the Maven Test worked. Need to take a closer look at this.